Fix: Units chaos
Added proper units to all values, while all measurements are now proper
objects. This is done on-fly, regardless of the state of the API.
Default data for gases, groups, linkers, metals updated. Fixes on all
tables (in different layout sizes) to support rendering of units.
Removed "size" property (aka length) from metal column, as this
measurement is not even valid.

All dashes have been replaced with proper dashes.
A few angstroms remains as the letter Å. This will be resolved in future
versions of this app.

We love unicode.