New: Actual API predictions!
The most important thing is that the app can now properly get actual
predictions from the API.

In less important news, a lot of things HAVE
to be changed in the UI in future updates to handle the limited data we
are getting out of it.

- The API returns only the diffusivity value. The UI was made to also
display data for diameter and permeability.
- Its unclear if the term Diffusion can be used in the UI instead of
diffusivity. To put it simply, diffusivity does not fit.
- Gas was designed to be an optional parameter, yet it is currently
required. This raises even more issues: Should we have a "default" Gas
for a "default" ZIF? Do we have a "default" ZIF now, or the one on which
we are defaulting was just a random choice. Should we pick a random
"default" gas? Why is choosing gas so important? (people won't read the
tutorial). How can we force the use to select a gas if they haven't
selected one?
- Actually, not selecting a gas is handled with JS: Do not select a gas
and press "Execution" (which to be honest needs to be renamed to
Execute). You will notice that the Gas modal appears out of nowhere. It
is much nicer on black theme, I promise. Yet, if the user closes the
modal via the "X" button (which to be honest shouldn't have existed if
gas is a requirement, they are returned to the previous: "I have no idea
why the app is not working" state of the app.
- If gas is not optional, then why is considered to be a Step 2? Why it
is so important to have the Gas as a step 2? Please explain in less than
160 characters, to update the subtitle on Step #2.
- Save, load & export functions have not been tested with the new
limited amount of parameters. We hope they work as intended. If not,
they will be fixed in a future update.
- There is currently no cache for API calls.
- There is currently absolutely no handling of errors. This is partially
intentional, as we are not quite sure how the app's Bootstrap will
interact with Wordpress Bootstrap and therefore it won't be fixed before
merging.

Enjoy!