Upd: Gas element UX behaviour
Gas is the only unit that is not pre-defined on the ZIF and therefore "none" is selected by default. Unfortunately, "none" is not accepted on the back-end and therefore the user has to be forced to make a valid choice before "Execution". To enforce this rule, the Gas modal is shown on "Execution" if a Gas was not selected. But the user can still select "none" as a Gas or simply dismiss the modal. This is an UX issue: As we don't want to alter the behaviour of the other modals (on which "none" is not an option) at this stage of the app's development, we are preventing the default hide.bs.modal behavior of the Gas modal to decide its fate.

Also:
- Minor changes on CSS to properly support a "warning" modal.
- Minor changes on CSS to properly support dark theme Bootstrap table backgrounds.