Upd: Restored Restore & Delete scenarios
Also fixed a bug which would have saved 2 scenarios with the exact same name. Really (if user has 1 scenario saved, which is Scenario 2, and tries to save another one, it would have also been named 2 as it is the 2nd scenario in storage - now that won't happen and all the names will be happy unique).