Upd: UX QoL improvements
Restored scenario is now highlighted and can not any longer be re-restored via the Restore button.
On the Intro Tab the 3rd step is highlighted if Execution is possible.
On the Intro Tab the "Execution" button highlights on demand the Execution button.
The header column on the "Select metal, linkers & groups" table is now hidden on <480px screens to finally allow at least some kind of usability on mobile devices.